Named some rng functions, also fixed s64/u64 types.

+++ include/types.h
@@ -1,16 +1,16 @@
 #ifndef _TYPES_H_
 #define _TYPES_H_
 
-typedef unsigned char  u8;
-typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
-typedef signed char    s8;
-typedef signed short   s16;
-typedef signed int     s32;
-typedef signed long    s64;
-typedef float          f32;
-typedef double         f64;
+typedef unsigned char      u8;
+typedef unsigned short     u16;
+typedef unsigned int       u32;
+typedef unsigned long long u64;
+typedef signed char        s8;
+typedef signed short       s16;
+typedef signed int         s32;
+typedef signed long long   s64;
+typedef float              f32;
+typedef double             f64;
 
 #ifndef NULL
     #define NULL 0
